Enable loading of lists and selection of word
The new lists format allows for an easier merge operation.
The word selection is very basic for now and only allows
for one review per word per reviewer.

Last but no least, the lists are only fetched once at this point,
and there is no cache invalidation. (This shouldn't be too difficult
to add later on, that's why I'm leaving it aside for now.)